Add alert when trying to run a program without enough credits

+++ se-engine/src/main/java/engine/logic/programData/program/Program.java
@@ -58,7 +58,7 @@ public interface Program extends Serializable {
     void addCreditCost(int creditCost);
     void incrementExecutionsCount();
     int getExecutionsCount();
-    int getAverageCreditCost();
+    long getAverageCreditCost();
 
     String getMainProgramNameOfThisProgram();
 }
